Autogenerate run command
