+++ heartbeat/cmd/root.go
@@ -1,9 +1,10 @@
 package cmd
 
 import cmd "github.com/elastic/beats/libbeat/cmd"
+import "github.com/elastic/beats/heartbeat/beater"
 
 // Name of this beat
 var Name = "heartbeat"
 
 // RootCmd to handle beats cli
-var RootCmd = cmd.GenRootCmd(Name, RunCmd)
+var RootCmd = cmd.GenRootCmd(Name, beater.New)
