+++ libbeat/cmd/root.go
@@ -3,13 +3,16 @@ package cmd
 import (
 	"flag"
 
+	"github.com/elastic/beats/libbeat/beat"
 	"github.com/spf13/cobra"
 )
 
 // GenRootCmd returns the root command to use for your beat. It takes
 // beat name as paramter, and also run command, which will be called if no args are
 // given (for backwards compatibility)
-func GenRootCmd(name string, runCmd *cobra.Command) *cobra.Command {
+func GenRootCmd(name string, beatCreator beat.Creator) *cobra.Command {
+	runCmd := genRunCmd(name, beatCreator)
+
 	rootCmd := &cobra.Command{
 		Use: name,
 		Run: runCmd.Run,
@@ -24,19 +27,6 @@ func GenRootCmd(name string, runCmd *cobra.Command) *cobra.Command {
 	rootCmd.PersistentFlags().AddGoFlag(flag.CommandLine.Lookup("path.logs"))
 	rootCmd.PersistentFlags().AddGoFlag(flag.CommandLine.Lookup("path.home"))
 	rootCmd.PersistentFlags().AddGoFlag(flag.CommandLine.Lookup("strict.perms"))
-
-	// Run subcommand flags, only available to *beat run
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("N"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("e"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("v"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("httpprof"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("cpuprofile"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("memprofile"))
-
-	// TODO deprecate in favor of subcommands (7.0):
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("configtest"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("setup"))
-	runCmd.Flags().AddGoFlag(flag.CommandLine.Lookup("version"))
 
 	// Inherit root flags from run command
 	// TODO deprecate when root command no longer executes run (7.0)
